% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Compute the growth rate}
\usage{
growth_rate(skw_t, skw_ini, date_t, date_ini = 0,
  method = c("exponential", "linear", "linear_std"))
}
\arguments{
\item{skw_t}{The weight at time t}

\item{skw_ini}{The weight at initial time}

\item{date_t}{The number day at time t}

\item{date_ini}{The number day at initial time}

\item{method}{The method to be employed. See the description to know the different methods}
}
\value{
The growth rate
}
\description{
The determination of the growth rate is a important parameter. Indeed, the growth rate can be used to compare different individuals. There are several methods to compute the growth rate.
}
\examples{

growth_rate(skw_t = 1.2, skw_ini = 1, date_t = 7, date_ini = 0,
method = "exponential")

# See \\url{https://econum.github.io/coral.growth/articles/} for more examples

}
