% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coral_growth}
\alias{coral_growth}
\title{Buoyant weight of corals over 56 days}
\format{
A data frame with 5 variables and 302 observations:
\describe{
\item{\code{id}}{A \strong{factor} with 10 levels.}
\item{\code{date}}{Number day.}
\item{\code{weight}}{Buoyant weight (in g).}
\item{\code{salinity}}{Salinity (in PSU).}
\item{\code{temperature}}{Temperature (in degree celsius).}
}
}
\usage{
coral_growth
}
\description{
Buoyant weight of corals over time. The measurements extend over 56 days.
}
\examples{
data(coral_growth)
class(coral_growth)
head(coral_growth)

}
\keyword{datasets}
